Rename instrumenters to _instrumenters to show it is private
