+++ scorep/_instrumenters/utils.py
